rename packages form kerberus to valhalla
