--- src/main/java/com/br/kerberus/urd/core/enumeration/HttpMethods.java
@@ -1,9 +1,0 @@
-package com.br.kerberus.urd.core.enumeration;
-
-public enum HttpMethods {
-    GET,
-    POST,
-    PUT,
-    PATCH,
-    DELETE
-}
