--- src/main/java/com/br/kerberus/urd/core/enumeration/LogLevel.java
@@ -1,6 +1,0 @@
-package com.br.kerberus.urd.core.enumeration;
-
-public enum LogLevel {
-
-    INFO, WARN, ERROR, DEBUG
-}
