--- src/main/java/com/br/kerberus/urd/core/enumeration/LogTimeType.java
@@ -1,7 +1,0 @@
-package com.br.kerberus.urd.core.enumeration;
-
-public enum LogTimeType {
-    MILLISECONDS,
-    SECONDS,
-    MINUTES
-}
