--- src/main/java/com/br/kerberus/urd/core/enumeration/LogType.java
@@ -1,13 +1,0 @@
-package com.br.kerberus.urd.core.enumeration;
-
-public enum LogType {
-
-    METHOD_CALL,
-    METHOD_RETURN,
-    HTTP_MESSAGE,
-    SOAP_MESSAGE,
-    REMOTE_CLIENT,
-    EXECUTION_TIME,
-    EXCEPTION,
-    GENERAL
-}
