--- src/main/java/com/br/kerberus/urd/exception/NoManagedException.java
@@ -1,7 +1,0 @@
-package com.br.kerberus.urd.exception;
-
-public class NoManagedException extends DomainException {
-    public NoManagedException(Exception e, String moreInfo) {
-        super(e.getMessage(), "Internal Server Error", 1, moreInfo);
-    }
-}
