--- src/main/java/com/br/kerberus/urd/service/LogService.java
@@ -1,10 +1,0 @@
-package com.br.kerberus.urd.service;
-
-import org.slf4j.Logger;
-
-public interface LogService {
-
-    Logger getLog();
-
-    void setLog(Logger log);
-}
