--- src/main/java/com/br/kerberus/urd/service/SoapLogService.java
@@ -1,19 +1,0 @@
-package com.br.kerberus.urd.service;
-
-import com.br.kerberus.urd.core.AspectLog;
-import org.aspectj.lang.annotation.Aspect;
-import org.slf4j.Logger;
-import org.springframework.stereotype.Component;
-
-@Aspect
-@Component
-public class SoapLogService extends AspectLog implements LogService{
-
-    private Logger log;
-
-    @Override
-    public Logger getLog() { return log; }
-
-    @Override
-    public void setLog(Logger log) {this.log = log;}
-}
