--- src/main/java/com/br/kerberus/urd/service/StartupApplicationLogService.java
@@ -1,35 +1,0 @@
-package com.br.kerberus.urd.service;
-
-import com.br.kerberus.urd.core.Application;
-import com.br.kerberus.urd.core.AspectLog;
-import com.br.kerberus.urd.core.SystemInformation;
-import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.annotation.Aspect;
-import org.aspectj.lang.annotation.Before;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-
-@Aspect
-@Component
-public class StartupApplicationLogService extends AspectLog implements LogService {
-
-    private Logger log;
-
-    @Override
-    public Logger getLog() { return log; }
-
-    @Override
-    public void setLog(Logger log) {this.log = log;}
-
-    public StartupApplicationLogService() {this.setLog(LoggerFactory.getLogger(StartupApplicationLogService.class)); }
-
-    @Before(value = "@annotation(com.br.kerberus.urd.annotation.LogMethodCall)")
-    public void logMethodCall(JoinPoint joinPoint) {
-
-        log.info("--------------------------------------------------------------------------------------------------");
-        log.info(String.format("Starting Application %s", "app"));
-        log.info(String.format("Information Application %s", new Application()));
-        log.info(String.format("System Information: %s", new SystemInformation().toString()));
-    }
-}
