+++ src/main/java/com/br/valhalla/urd/config/UrdConfig.java
@@ -1,0 +1,11 @@
+package com.br.valhalla.urd.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.EnableAspectJAutoProxy;
+
+@Configuration
+@EnableAspectJAutoProxy(proxyTargetClass = true)
+@ComponentScan(basePackages="com.br.kerberus.urd")
+public class UrdConfig {
+}
