+++ src/main/java/com/br/valhalla/urd/core/enumeration/HttpMethods.java
@@ -1,0 +1,9 @@
+package com.br.valhalla.urd.core.enumeration;
+
+public enum HttpMethods {
+    GET,
+    POST,
+    PUT,
+    PATCH,
+    DELETE
+}
