+++ src/main/java/com/br/valhalla/urd/core/enumeration/LogLevel.java
@@ -1,0 +1,6 @@
+package com.br.valhalla.urd.core.enumeration;
+
+public enum LogLevel {
+
+    INFO, WARN, ERROR, DEBUG
+}
