+++ src/main/java/com/br/valhalla/urd/core/enumeration/LogTimeType.java
@@ -1,0 +1,7 @@
+package com.br.valhalla.urd.core.enumeration;
+
+public enum LogTimeType {
+    MILLISECONDS,
+    SECONDS,
+    MINUTES
+}
