+++ src/main/java/com/br/valhalla/urd/core/enumeration/LogType.java
@@ -1,0 +1,13 @@
+package com.br.valhalla.urd.core.enumeration;
+
+public enum LogType {
+
+    METHOD_CALL,
+    METHOD_RETURN,
+    HTTP_MESSAGE,
+    SOAP_MESSAGE,
+    REMOTE_CLIENT,
+    EXECUTION_TIME,
+    EXCEPTION,
+    GENERAL
+}
