+++ src/main/java/com/br/valhalla/urd/exception/ManagedException.java
@@ -1,0 +1,7 @@
+package com.br.valhalla.urd.exception;
+
+public class ManagedException extends DomainException {
+	public ManagedException(String debugMessage, String userMessage, int errorCod, String moreInfo) {
+		super(debugMessage, userMessage, errorCod, moreInfo);
+	}
+}
