+++ src/main/java/com/br/valhalla/urd/exception/NoManagedException.java
@@ -1,0 +1,7 @@
+package com.br.valhalla.urd.exception;
+
+public class NoManagedException extends DomainException {
+    public NoManagedException(Exception e, String moreInfo) {
+        super(e.getMessage(), "Internal Server Error", 1, moreInfo);
+    }
+}
