+++ src/main/java/com/br/valhalla/urd/service/LogService.java
@@ -1,0 +1,10 @@
+package com.br.valhalla.urd.service;
+
+import org.slf4j.Logger;
+
+public interface LogService {
+
+    Logger getLog();
+
+    void setLog(Logger log);
+}
